Add list models with public and like data
